fix(pre-commit): remove pip-tools and bandit, replace flake8 in favor of flakehell
pip-tools generates different results in the CI than in the development
environment breaking the CI without an easy way to fix it. Therefore it
should be run by the developers periodically.

bandit takes a lot of time to run, so it slows down too much the
commiting, therefore it should be run only in the CI.

flakehell has more features than flake8