ci(github): add bandit and black tests
ci(pre-commit): tweak mypy, safety and flakehell configuration

To prevent undesired tests to be run.

chore(makefile): change order of dependency update

The development requirements now depend on the documentation ones.

fix(makefile): correct the files to test

Particularize the template for the cookiecutter case, such as:

* Removing the coveralls report because we can't test the hooks due to
integration with thirds, we should mock them but I'm not up to it for
now, if you do, please do a PR. And the rest of the tests check the
output of the cookiecutter, not the program lines, so a coverage report
makes no sense in this case.

docs: create mkdocs documentation

fix: correct the generation of ssh keys for ci deployment

Each repository needs it's own ssh key, so we need to generate it
in the post hooks.

chore: clean up commented code to generate virtualenvs

feat(post-hook): configure Github secrets

tests: make mypy and flakehell compatible

tests: add autoimport to manage imports

tests: build dependencies and import the generated package

chore: update dependencies

chore(makefile): add version command

fix(setup.py): improve version extraction

We used exec before, which opens the door for vulnerabilities.

docs: add docstrings and typing to all the code